Implement Sprint 3
* Implement GetTable operator

* [WIP] Begin table_scan implementation

* Fix table scan on execute function call to be able to compile

* Implement ReferenceSegment

* Fix linting and formatting errors

* refactored most comments of last review and some additional IDE comments

* Implement table scan type dispatching

* Implement table scan segment dispatch

* Implement Table::emplace_chunk

* Apply emplace_chunk method

* Implement dispatched comparator

* Implement value scan

* Implement dictionary segment scan and adapt function defenitions

* continuing table scan and abstract operator

* Implement add_column_definition function and enable tests

* Use coorect referenced table

* Fix dictionary segment debug assert

* hopefully implemented type checking for tablescan execution

* Include missing dependency

* Implement _scan_reference_segment

* table scan type checking working

* all tests passed

* formatting and linting

* started commenting

* Remove dispatched_comparator function by dispatching the comparator once

* Format and lint table_scan

* Add comments to table_scan functions

* small comment

* small assert

* small fixes

* Add comment for nested if construct

+++ src/lib/storage/chunk.cpp
@@ -26,11 +26,11 @@ void Chunk::append(const std::vector<AllTypeVariant>& values) {
 
 std::shared_ptr<BaseSegment> Chunk::get_segment(ColumnID column_id) const { return _segments[column_id]; }
 
-uint16_t Chunk::column_count() const { return _segments.size(); }
+uint16_t Chunk::column_count() const { return static_cast<uint16_t>(_segments.size()); }
 
 uint32_t Chunk::size() const {
   if (column_count() > 0) {
-    return _segments[0]->size();
+    return static_cast<uint32_t>(_segments[0]->size());
   }
   return 0;
 }
